+++ backend/src/main/java/calculator/Divides.java
@@ -18,24 +18,10 @@ public final class Divides extends Operation
    *
    * @param elist The list of Expressions to divide
    * @throws IllegalConstruction    If an empty list of expressions if passed as parameter
-   * @see #Divides(List<Expression>,Notation)
+   * @see #Divides(List<Expression>)
    */
-  public /*constructor*/ Divides(List<Expression> elist) throws IllegalConstruction {
-	this(elist, null);
-  }
-
-    /**
-     * Class constructor specifying a number of Expressions to divide,
-     * as well as the notation used to represent the operation.
-     *
-     * @param elist The list of Expressions to divide
-     * @param n The Notation to be used to represent the operation
-     * @throws IllegalConstruction  If an empty list of expressions if passed as parameter
-     * @see #Divides(List<Expression>)
-     * @see Operation#Operation(List<Expression>,Notation)
-     */
-  public Divides(List<Expression> elist, Notation n) throws IllegalConstruction {
-	super(elist,n);
+  public Divides(List<Expression> elist) throws IllegalConstruction {
+	super(elist);
 	symbol = "/";
 	neutral = 1;
   }