+++ backend/src/main/java/calculator/Operation.java
@@ -1,10 +1,10 @@
 package calculator;
 
+import visitor.Printer;
 import visitor.Visitor;
 
 import java.util.ArrayList;
 import java.util.List;
-import java.util.stream.Stream;
 
 /**
  * Operation is an abstract class that represents arithmetic operations,
@@ -30,32 +30,13 @@ public abstract class Operation implements Expression
    */
   protected int neutral;
 
-  /**
-   * The notation used to render operations as strings.
-   * By default, the infix notation will be used.
-   */
-  public Notation notation = Notation.INFIX;
-
-  /** It is not allowed to construct an operation with a null list of expressions.
-   * Note that it is allowed to have an EMPTY list of arguments.
-   *
-   * @param elist	The list of expressions passed as argument to the arithmetic operation
-   * @throws IllegalConstruction	Exception thrown if a null list of expressions is passed as argument
-   */
-  protected /*constructor*/ Operation(List<Expression> elist)
-		  throws IllegalConstruction
-	{
-		this(elist, null);
-    }
-
 	/** To construct an operation with a list of expressions as arguments,
 	 * as well as the Notation used to represent the operation.
 	 *
 	 * @param elist	The list of expressions passed as argument to the arithmetic operation
-	 * @param n 	The notation to be used to represent the operation
 	 * @throws IllegalConstruction	Exception thrown if a null list of expressions is passed as argument
 	 */
-	protected /*constructor*/ Operation(List<Expression> elist,Notation n)
+	protected /*constructor*/ Operation(List<Expression> elist)
 			throws IllegalConstruction
 	{
 		if (elist == null) {
@@ -63,7 +44,6 @@ public abstract class Operation implements Expression
 		else {
 			args = new ArrayList<>(elist);
 		}
-		if (n!=null) notation = n;
 	}
 
 	/**
@@ -74,6 +54,15 @@ public abstract class Operation implements Expression
 	public List<Expression> getArgs() {
   	return args;
   }
+
+	/**
+	 * getter method to return the symbol of the arithmetic operation.
+	 *
+	 * @return	The symbol of the arithmetic operation (e.g. "+", "-", "*", "/").
+	 */
+	public String getSymbol() {
+		return symbol;
+	}
 
 	/**
 	 * Abstract method representing the actual binary arithmetic operation to compute
@@ -100,79 +89,33 @@ public abstract class Operation implements Expression
 	 * @param v	The visitor object
 	 */
   public void accept(Visitor v) {
-  	for(Expression a:args) { a.accept(v); }
   	v.visit(this);
-  }
-
-	/**
-	 * Count the depth of an arithmetic expression recursively,
-	 * using Java 8 functional programming capabilities (streams, maps, etc...)
-	 *
- 	 * @return	The depth of the arithmetic expression being traversed
-	 */
-	public final int countDepth() {
-	    // use of Java 8 functional programming capabilities
-	return 1 + args.stream()
-			   .mapToInt(Expression::countDepth)
-			   .max()
-			   .getAsInt();  
-  }
-
-	/**
-	 * Count the number of operations contained in an arithmetic expression recursively,
-	 * using Java 8 functional programming capabilities (streams, maps, etc...)
-	 *
-	 * @return	The number of operations contained in an arithmetic expression being traversed
-	 */
-	public final int countOps() {
-	    // use of Java 8 functional programming capabilities
-	return 1 + args.stream()
-			   .mapToInt(Expression::countOps)
-			   .reduce(Integer::sum)
-			   .getAsInt();
-  }
-
-  public final int countNbs() {
-	    // use of Java 8 functional programming capabilities
-	return args.stream()
-			   .mapToInt(Expression::countNbs)
-			   .reduce(Integer::sum)
-			   .getAsInt();  
   }
 
   /**
    * Convert the arithmetic operation into a String to allow it to be printed,
-   * using the default notation (prefix, infix or postfix) that is specified in some variable.
+   * using the default printer
    *
    * @return	The String that is the result of the conversion.
    */
   @Override
   public final String toString() {
-  	return toString(notation);
+		Printer p = new Printer();
+		this.accept(p);
+		return p.getResult();
   }
 
   /**
-   * Convert the arithmetic operation into a String to allow it to be printed,
+   * Appeal to the visitor to convert the arithmetic operation into a String to allow it to be printed,
    * using the notation n (prefix, infix or postfix) that is specified as a parameter.
    *
    * @param n	The notation to be used for representing the operation (prefix, infix or postfix)
    * @return	The String that is the result of the conversion.
    */
   public final String toString(Notation n) {
-	   Stream<String> s = args.stream().map(Object::toString);
-	   return switch (n) {
-		   case INFIX -> "( " +
-				   s.reduce((s1, s2) -> s1 + " " + symbol + " " + s2).get() +
-				   " )";
-		   case PREFIX -> symbol + " " +
-				   "(" +
-				   s.reduce((s1, s2) -> s1 + ", " + s2).get() +
-				   ")";
-		   case POSTFIX -> "(" +
-				   s.reduce((s1, s2) -> s1 + ", " + s2).get() +
-				   ")" +
-				   " " + symbol;
-	   };
+	   Printer p = new Printer(n);
+	   this.accept(p);
+	   return p.getResult();
   }
 
 	/**