+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 ## [Unreleased]
 
 ### Added
+
+### Changed
+
+## Fixed
+- In case of a usage of several notations inside an expression (infix, prefix and postfix), the expression is considered as the type of the top-level.
+- Removal of the notation field in the class Operation. Now, the notation is only used when we use the Printer Visitor.
+
+
+
+## [v0.0.1]
+
+### Added
 - Initial project structure based on `calculator-cucumber-2026` template.
 - Basic arithmetic operations: Addition, Subtraction, Multiplication, Division.
 - Support for Prefix, Infix, and Postfix notations.
@@ -19,8 +31,5 @@ All notable changes to this project will be documented in this file.
 
 ### Changed
 - Updated `.gitignore` to ignore the `target/` directory and others files.
-- Updated dependencies of Junit and Cucumner to the latest version (6.0.2 and 7.34.2 respectively).
+- Updated dependencies of Junit and Cucumber to the latest version (6.0.2 and 7.34.2 respectively).
 
-
-## Fixed
--